requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
click==8.3.1
distro==1.9.0
fastapi==0.128.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
openai==2.14.0
pydantic==2.12.5
pydantic-core==2.41.5
sniffio==1.3.1
starlette==0.50.0
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.40.0
mcp[cli]==1.10.1
openai==2.14.0
python-dotenv
ipykernel
httpx
